+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.7.0 - TBD
+
+### Added
+
+- [#26](https://github.com/zendframework/zend-navigation/pull/26) adds:
+  - `Zend\Navigation\View\ViewHelperManagerDelegatorFactory`, which decorates
+    the `ViewHelperManager` service to configure it using
+    `Zend\Navigation\View\HelperConfig`.
+  - `ConfigProvider`, which maps the default navigation factory and the
+    navigation abstract factory, as well as the navigation view helper.
+  - `Module`, which does the same as the above, but for zend-mvc
+    applications.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
 ## 2.6.2 - TBD
 
 ### Added